src/mlt/pyproject.toml
[project]
name = "mlt-toolbox"
version = "0.1.0"
description = "ML Toolbox Remote Helper - LSP proxy and container utilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
mlt = "mlt.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mlt"]
